+++ tests/test_cookie.py
@@ -26,7 +26,6 @@ def test_cookiecutter(cookies, monkeypatch):
     assert CookieCutter.CLUSTER_NAME == ""
     assert CookieCutter.CLUSTER_CONFIG == ""
     assert CookieCutter.get_cluster_option() == ""
-    assert CookieCutter.get_advanced_argument_conversion() is False
     sys.modules.pop("CookieCutter")
 
 
@@ -34,13 +33,11 @@ def test_cookiecutter_extra_context(cookies, monkeypatch):
     result = cookies.bake(template=str(pytest.cookie_template),
                           extra_context={"sbatch_defaults": "account=foo",
                                          "cluster_name": "dusk",
-                                         "cluster_config": "slurm.yaml",
-                                         "advanced_argument_conversion": "yes"})
+                                         "cluster_config": "slurm.yaml"})
     monkeypatch.syspath_prepend(str(result.project_path))
     from CookieCutter import CookieCutter
     assert CookieCutter.SBATCH_DEFAULTS == "account=foo"
     assert CookieCutter.CLUSTER_NAME == "dusk"
     assert CookieCutter.CLUSTER_CONFIG == "slurm.yaml"
     assert CookieCutter.get_cluster_option() == "--cluster=dusk"
-    assert CookieCutter.get_advanced_argument_conversion() is True
     sys.modules.pop("CookieCutter")